requirements.txt
# ─────────────────────────────────────────────────────────────
# SentiSign requirements.txt
# Install with:  pip install -r requirements.txt
# Python 3.10 recommended
# ─────────────────────────────────────────────────────────────

# TTS — Chatterbox (real emotion control via exaggeration + cfg_weight)
chatterbox-tts
elevenlabs
python-dotenv

# Sentence generation — flan-t5-large
# torch + transformers already pulled in by chatterbox-tts
sentencepiece          # REQUIRED by T5Tokenizer — do not skip
accelerate             # speeds up flan-t5-large on GPU

# Audio playback — no SOX, no system dependencies
sounddevice
scipy
# mediapipe is not compatible with numpy 2.x (ABI break)
numpy<2.0
opencv-python==4.8.1.78
mediapipe==0.10.9
torchvision==0.21.0
onnx==1.14.1
protobuf==3.20.3
scikit-learn
pandas
pillow
transformers
fastapi
uvicorn[standard]
python-multipart
